remove test file and add test code into translator.rs
