+++ lib/saluki-components/src/sources/otlp/logs/mod.rs
@@ -1,5 +1,2 @@
 pub mod transform;
 pub mod translator;
-
-#[cfg(test)]
-mod translator_test;
